Updated openstack/openstack Project: openstack/tempest  9dd89404d3c1f867200ad6b8fc90d304d200bcf8
Adding the ability to use pre-created ports in Network scenarios

It's possible to launch an instance with virtual interfaces
either by specifying the network the interface should be attached
to or the port.
There are cases (e.g. SR-IOV) where it's only possible to launch
an instance with a pre-created port.
Enabling the creation of  a port with known vnic_types (Currently
'normal','direct' and 'macvtap') and launching an instance with this port
attached.

Nova support Neutron SR-IOV ports https://wiki.openstack.org/wiki/Nova-neutron-sriov

Change-Id: Id5cbd50d285ac0c7e8099c30151ace9d99f02020