Updated openstack/openstack Project: openstack/horizon  74e8d1b34a94cd867174c6c8c2717da798d27d7e
Add virtual size to image detail

In Icehouse, Glance added a new field, 'virtual_size', to hold the
virtual size (as opposed to on-disk size) of sparsed or compressed
images. In most cases, this is far more valuable to know than the
on-disk size of an image; for instance, the official Fedora 21 QCOW2
image is only 150 Mb on disk, but is a 3 Gb image. This patch adds an
optional 'virtual size' field to the image detail view, and makes the
"Launch Instance" modal use virtual size if it's available.

Closes-Bug: 1420944
Change-Id: I072635781c56f81502586afd69fe6691bd264a92