Updated openstack/openstack Project: openstack/nova  24bb0f981cdd526fc3592fe712e31b98e26c6bf6
Make obj_set_defaults() more useful

This makes obj_set_defaults() not overwrite existing values set on the object
with the default. The idea of obj_set_defaults() is to allow filling in fields
with default values if they are otherwise unknown. Before this patch, the
following code would not work, but does after:

 obj = MyObject(foo=123)
 obj.obj_set_defaults()
 # obj.foo would be overwritten with the default

Related to blueprint kilo-objects

Change-Id: I921f37e6bd8fe9c709066b2a75bcad94255a86dc