Updated openstack/openstack Project: openstack/python-keystoneclient  ae784225627f1652638edc0bf273a41519a438d5
Ignore all failures removing catalog when logging token

Operations could fail if the response was logged and had a 'token'
field that's not a dict. The fix is to ignore all errors when
trying to remove the service catalog from the response.

Also, enhanced the service catalog removal code to support V2
tokens.

Closes-Bug: 1420080
Change-Id: I35b971415744825e8e5f00f30dcf193d04ee699a