Updated openstack/openstack Project: openstack/nova  ceb9ef90168c73c9a29b0e98043bcb85f27675a1
libvirt: remove _destroy_instance_files shim

The instance parameter passed to _destroy_instance_files is a nova
object now so we can remove the shim and todo and move that code up to
the cleanup method. A note is added to clarify the usage of the
'clean_attempts' system_metadata.

The cleanup method is called by destroy and _create_domain_and_network
which are also getting an instance object.

Also changes one remaining place in _create_domain_and_network where the
instance.uuid field was accessed like a dict (changed to use dot
notation to re-enforce that it's an object).

Partial-Bug: #1420322

Change-Id: Ia4d204df233fd40a9008e6cc010446c62e18a3b9