Updated openstack/openstack Project: openstack/heat  888970da888dda683e3ddb24cd3bed708b19ba63
Use domain_id from auth token when not configured

When the stack_user_domain_admin_user/password are used to authenticate,
the domain_id they're included in is returned in the token. If the user
has configured Heat using a stack_user_domain_name then the ID will be
needed to create stack users. Previously we used the domain list
functionality in keystoneclient, but this required admin level
privileges in the default keystone policy file. There is no need to list
though, as we already have the domain ID.

Change-Id: I0eaad62502ba55bb52bc95192318f3f12a95062b
Closes-Bug: #1415136