Updated openstack/openstack Project: openstack/keystone  ff9778194a2e805b4b65dc6959598458eb394213
Improve creation of expected assignments in tests

Creation of expected role assignment entities and
links are in separate methods. The advantage is
to only create the link on methods that do not
need the entity. Test references are fixed.

In addition, expected assignment link is get from
the expected entity, instead of from a separate
attribute on assert methods at test_v3.

Partial-Bug: 1406640

Change-Id: I7c81dcf2627f2bfb712591cb33abd6fd03a3a52a