Updated openstack/openstack Project: openstack/keystone  259adbb6754eab634d1b37b7fb50211aadabe553
Log exceptions safely

The Python standard library logging module expects to be given a string
as the first argument to all of the log methods. If it receives
something that is not a string, it tries to convert it to a string. The
old logging ContextAdapter dealt with this transparently by using
unicode(), but oslo.log on longer uses the ContextAdapter. Since the
underlying code is using the equivalent of str(), we need to pass real
string objects in to the log calls, instead of passing just exception
objects.

Change-Id: I0bc3b1e9761441178cb5400559c6a84a6fb33f06