Updated openstack/openstack Project: openstack/keystone  5291236a53a32ecceaf254406bdfacf5d2d7800e
Fix evaluation logic of federation mapping rules

In the evaluation of federation mapping rules, when "regex" is true,
it's only checked that if the values from assertion match the values
from mapping rules, and "any_one_of" and "not_any_of" options are
bypassed. So if one specifies "regex: True" and "not_any_of" at the
same time, he will got an unexpected result that assertion with values
in "not_any_of" can pass the evaluation.

The expected behaviour, when "regex" is true, should be matching values
in assertion and mapping rules using regular expression, if match with
"any_one_of" or not match with "not_any_of", pass the evaluation,
otherwise fail the evaluation.

Change-Id: Ic6969c6dc23cff3abce775711f9ed01ffdf8dcb1
Closes-Bug: #1414961