Updated openstack/openstack Project: openstack/keystone  92233ef100c46bcf739d9904d9de21cb237e7e26
Don't try to convert LDAP attributes to boolean

Currently, in the ldap2py function, attribute values are attempted
to be converted to python friendly types.

It attempts to convert an attribute to boolean first, then fallback
as string. Any attribute value that is equal to 'TRUE' or 'FALSE'
are converted to boolean value.

Only 'enabled' attribute require this type of handling, and it is
already by enabled2py function. The additional handling on ldap2py
function should be removed.

Change-Id: Ifd1e91f98ae22a5e5f7b9b2ddbfddba156e16a6d
Closes-Bug: #1411478