Updated openstack/openstack Project: openstack/nova  e61723715eb33f87f432dd8472e4e19f99d026eb
patch out nova libvirt driver event thread in tests

The nova libvirt driver spawns a native thread to listen to events
coming from libvirt. This thread is spawned on driver
initialization. During testing we initial the libvirt driver quite a
lot, which means we've got extra threads flapping around during
tests. Even though these mostly are in an idle sleep loop, we have
tests that ensure sleeps happen in a particular order, called a
particular way (to ensure retry/fallback happens as expected). Which
means these threads are a source of race conditions in the tests.

The following adds a disable_event_loop call in fakelibvirt which can
either be called explicitly, or will always be used as part of the
fakelibvirt fixture.

To ensure this is doing what we think, this also adds a
PoisonFunctions fixture which does a global patch of the problematic
function to a sys.exit. So if any code calls that function without
locally patching it, the test dies. A sys.exit was used instead of an
exception to ensure we bypass any error handling, as touching this
function without patching is something we definitely want to avoid.

Change-Id: I813ea6abe168d3ae3b3b22fac1c6772615235eaa