Updated openstack/openstack Project: openstack/swift  af2ec5c634d0b6b83b72252d11ed1c5a5b31652d
Make container GET call authorize when account not found

When an account was not found, ContainerController would
return 404 unconditionally for a container GET or HEAD request,
without checking that the request was authorized.

This patch modifies the GETorHEAD method to first call any
callback method registered under 'swift.authorize' in the
request environ and prefer any response from that over the 404.

Closes-Bug: 1415957

Change-Id: I4f41fd9e445238e14af74b6208885d83698cc08d