Updated openstack/openstack Project: openstack/python-manilaclient  74325af4b7bc3c4f1bdea78a90c43e339be88908
Add Manila client and CLI support for listing scheduler pools

The pools feature of Manila core added an API to list existing
pools, needed by a few admin-only features such as manage/unmanage.
But like Cinder, no CLI command was added to provide easy access
to the feature.  This is rectified in this patch with a pool-list
command and associated unit and functional tests.

Closes-Bug: #1422043
Change-Id: I4613feafa4ca79400a7419be488252b3324cc231