Updated openstack/openstack Project: openstack/python-troveclient  51aa1a8dbd1b45e6abdaf571d1d73a6d7405fd3a
Pass all kwargs through to adapter

Part of the benefit of clients using the keystoneclient adapter to
manage the session object is that we can have a standard set of kwargs
that can be controlled from the adapter. This only works if the unknown
kwargs are passed through to the adapter.

Change-Id: I78f195ef739dbfe33a08f4dc1922b14ec77ae833