Updated openstack/openstack Project: openstack/nova  1c5659f036cbaafbe313cb5042eb53f334541a6a
Libvirt: vcpu_model support

Utilize the virt_cpu_model in libvirt for guest cpu config.

Currently we try to update the instance's cpu config each time we
recreated the xml, which means if there are some host configuration changes,
the changes will be applied to the instance as well.

This is same as current implementation, but we may need some enhancement
there.

Implements: blueprint resource-objects
Change-Id: I95f52ed2dc37d6d261a9a8ebcb59aa53c3c32f65