Updated openstack/openstack Project: openstack/trove-integration  ef6f6ef72ea1b481f83669a6aceb2143635ded17
Create DIB elements for HP Vertica on Ubuntu

A specification for this change was submitted for review in
https://review.openstack.org/#/c/151126/

Changes:
- Creates the disk-image-builder elements for
  installing HP Vertica on Ubuntu.

- Introduces new env variables DATASTORE_PKG_LOCATION which
  specifies location to look for package installer file.

- It would be the job of elements(extra-data.d) to discover
  existence of DATASTORE_PKG_LOCATION, and notify appropriate
  error message in case of non-existent DATASTORE_PKG_LOCATION.

To execute kick-start for Vertica, user has to download
Vertica package from https://my.vertica.com and place
file at DATASTORE_PKG_LOCATION and update the value of
DATASTORE_PKG_LOCATION to point to actual package-file.

For eg.
if package file-name is vertica.deb and it is placed in /tmp,
then to run kick-start proceed as follows:

  $ export DATASTORE_PKG_LOCATION=/tmp/vertica.deb
  $ ./redstack kick-start vertica

Partially Implements: blueprint vertica-db-support

Change-Id: Ibc64b42eb5d2f4722eb27a7bc730dda98014347f