Updated openstack/openstack Project: openstack/nova  bfca2529166937803b86ba34a4a51c28f2a9555f
libvirt: Fix logically inconsistent host NUMA topology

The _fake_caps_numa_topology() sets up a fake host NUMA topology
that the tests will use. The definition it has setup though is
logically inconsistent. The setup as 4 nodes with 2 CPUs in each
cell, but it then says there are 2 cores per socket and 2 thread
siblings per core. This implies 4 CPUs per cell, not 2. Either
the topology should have 2 cores with 1 thread, or 1 core with
2 threads. The test cases using this topology are written to
treat it as meaning 1 core with 2 threads, so fix it according
to this interpretation

Related-bug: 1417671
Change-Id: I9d7928a6f9b15f8eb29ef92606e6ca1d2b688a6e