Updated openstack/openstack Project: openstack/nova  69f4b44691ddabc2cfa8c08539a51d255646e173
libvirt: fix emulator thread pinning when doing strict CPU pinning

When guest vCPUs are confined to particular set of host pCPUs, the
emulator threads from QEMU are intended to be pinned to the union
of the host pCPUs that the vCPUs are associated with.

The code in Nova was in fact confining the emulator threads to the
union of pCPUs in the host NUMA nodes that the guest was confined
to. When running guests in NUMA mode with shared CPUs / overcommit
this was functionally identical, but when running guests with
dedicated CPUs this was incorrect.

The (incorrect) libvirt config being generated was

  <cputune>
      <shares>4096</shares>
      <vcpupin vcpu='0' cpuset='0'/>
      <vcpupin vcpu='1' cpuset='1'/>
      <vcpupin vcpu='2' cpuset='4'/>
      <vcpupin vcpu='3' cpuset='5'/>
      <emulatorpin cpuset='0-5'/>
  </cputune>

when it should have been

  <cputune>
      <shares>4096</shares>
      <vcpupin vcpu='0' cpuset='0'/>
      <vcpupin vcpu='1' cpuset='1'/>
      <vcpupin vcpu='2' cpuset='4'/>
      <vcpupin vcpu='3' cpuset='5'/>
      <emulatorpin cpuset='0-1,4-5'/>
  </cputune>

This was not caught because the unit tests were only checking with
a host numa topology that had the same number of cpus as the guest
topology. The test is changed to make the host topology massive
compared to guest topology

Closes-bug: 1417671
Change-Id: I81e0fa1d9b09ec2df2af5e21c7d95b21be435f90