Updated openstack/openstack Project: openstack/nova  3f33e76d19b3945bef44ce845a5eb2a430a1fa40
Implement api samples test for os-baremetal-nodes Part 2

Implements an API sample test for a GET request for os-baremetal-nodes
which was previously manually generated and wrong. Removes api samples
for add-interface, create, and remove-interface which were removed
for os-baremetal-nodes at some point.

This should bring the os-baremetal-nodes plugin in sync with reality
for both v2 and v2.1

Partially implements blueprint v2-on-v3-api

Change-Id: Ibd55b43368485b7f460ef6564f611df136275562