Updated openstack/openstack Project: openstack/keystone  142ece37dc2dfc8963524a6991c58cf8f00b094d
fix a small issue in test_v3_auth.py

In test_project_id_scoped_token_with_user_id_401 method, we generate
an useless project_id because the project_id parameter wouldn't be
used when creating project. And also we'd better pass a project_id
that equals to the id in project ref.

Change-Id: I6ebe7f7d86797fa88264d73d339bd1a108091a16