Updated openstack/openstack Project: openstack/nova  6a4c1a4c41e5be5538fa285531e3618d4ad34e2d
Cells: Record initial database split in devref

In moving forward with cells v2 the existing database tables need to be
divided into those that belong at the API level and those that belong
within a cell.  This will be an ongoing and evolving discussion but the
decisions made to this point should be documented.

Change-Id: Ie888d22c8d08173cc86f5b4068657a5224cad34a