Updated openstack/openstack Project: openstack/keystone  ba10e4bee22b8bf798c448e82cb30977f5e5322b
Add a domain to federated users

Currently in the token returned from keystone, for a federated
user, the 'domain' section is missing. This was as designed, since
federated users do not exist in keystone. However this was caused
some issues with the upcoming non-persistent tokens, as well as
keystonemiddleware.

The solution should be to add a domain section, where we use a
reserved value of 'Federated', for both ID and name of the domain.

Change-Id: I4ad9fb55e8258e091ac78f628e5cb5b3add611a3
Implements bp federated-direct-user-mapping