Updated openstack/openstack Project: openstack/horizon  47f230fa6338ded740de504c2421f36f8d6fb299
Revert "Fix shared firewall policy can not be displayed in horizon"

This reverts commit ca8df08bb8ed03981b7a0f9d6fedb3418b061804.

The reverted commit introduced a regression that all firewall
related resources from all projects are shown in the project panel
if a user has admin role. It violates the design policy of the
Project dashboard (which focuses on project-related resources).
A fix for the original bug 1294541 is coming in the subsequent patch.

Closes-Bug: #1422081
Related-Bug: #1294541
Change-Id: Ie8864b92b4ffb59681c0d18ac8b5f09f636d5fba