Updated openstack/openstack Project: openstack-infra/devstack-gate  57be7c262e521a04cf6a2733eba311075f013a91
actually save hook logs

We've not been saving the hook logs because we are attempting to do
the save without enough permissions.

example:

http://logs.openstack.org/94/156594/1/check/check-swift-dsvm-functional/310e7a6//console.html#_2015-02-17_14_38_21_451

All other logs are moved with sudo, we should do that here as well.

For test purposes we make SUDO overridable for this call. In the
future we can consider what other SUDO calls are possibly overridable.

Change-Id: I6d4a378edd65cbe54243f8089ed9d0bc2b1a6478