Updated openstack/openstack Project: openstack-infra/zuul  79427a40201e190c4162e98817695adb0a8fd337
Stop metajob application of undefined attributes

Previously, metajobs set defaults for boolean attributes that were not
defined, and copied to them to matched jobs.  This could result in a
metajob that did not define a voting attribute unintentially
overriding a matched job with the default voting value (True).  This
change ensures that metajobs set values only for explicitly supplied
boolean attributes.

Change-Id: I99dc3a58c63946ddb54849b1629101b062159a69
Closes-Bug: #2000166