Updated openstack/openstack Project: openstack/nova  9aad8622f705e64d4fdb1499a4ed7a2b63a10f04
Don't mock time.sleep with None

Patching time.sleep with 'None' is always the wrong
thing to do, because it removes the ability for the
thread scheduler to do anything useful, changes the
semantics of the system-under-test and may lead to
deadlocks. Patch with eventlet.sleep(0) instead.

See Dan Berrange's analysis in comment #2 at
https://bugs.launchpad.net/nova/+bug/1412742

Change-Id: Iad6cb63cb579bb9e0d8ef0c6e626c80053468157
related-bug: #1412742