Updated openstack/openstack Project: openstack/neutron  dd5c4ac517c087e7cbc8480f39376e40461e4ac8
Fix FIP agent gw port delete based on external net

Today the FIP agent gateway port for DVR is deleted
based on the host. When there are multiple external
networks, then the port deletion for the second
external network may fail.

So the current fix checks for the valid host and
external network id and then deletes the FIP agent
gw port if it is the last one to be deleted.

Closes-Bug: #1421886

Change-Id: Ic5dfb88409c39d06d912263b338d0c5f7ebd3bcb