Updated openstack/openstack Project: openstack/nova  53ae5da0fa424479da5695b03f44420bef775969
Libvirt: Gracefully Handle Destroy Error For LXC

When libvirt destroys a container it will attempt to send a shutdown
to the init process, then kill all of the procceses left over. If
they don't die fast enough, libvirt will raise an error. The
container will eventually die in most cases. So, instead of erroring
the destroy, we should let the wait_for_destroy logic continue to
see if the instance goes away.

Change-Id: I1feb767e82bc77ad7ae63852a043d67844245626
Closes-Bug: #1413705