Updated openstack/openstack Project: openstack/nova  82357ccb6b912ac2737dd2d3d3bc99f8bf769e1e
Expand try/except for get_machine_ips

Commit b838ca28e0ffda21b298a4514a647fc74c154f4a added get_machine_ips
and copied it from swift but didn't put the try/except in the same
scope, essentially swift commit 76afb46d91bb32dfcbda829408d6b0735526b92f.
This causes a ValueError failure race bug in resize tests in the gate.

This change simply expands the scope of the try/except like it is in
swift and adds a test to cover that.

Note that we should move this utility function into oslo at some point
if multiple projects are needing this.

Closes-Bug: #1422901

Change-Id: Iccceca4b9706ac7723563fbd2065f40586400527