Updated openstack/openstack Project: openstack/nova  fcb4ee62ebf249b49fce041dd46e91444e7f11f0
Refactor how to remove compute nodes when service is deleted

service_id is about to be deprecated and won't be updated in a next patch.
We need instead to make sure that we can remove the nested compute nodes
by filtering on the host field.

Note: in order to get the host field value, we need to do an extra DB call
but we assume it as reasonable as it's only an admin action for a very
specific case (service deletion)

Partially-Implements: blueprint detach-service-from-computenode

Change-Id: Iaf03c45b23f2b3fd759f8e2a0deb92c0bea2a408