Updated openstack/openstack Project: openstack/taskflow  67905d59f6faf93aaa887416dc67416f237614a3
Improve multilock class and its associated unit test

In the multilock unit test check that all values were
acquired (as expected). Also use a class constant for
the number of threads to spin up when testing timing
overlaps and use the module constant for the duration
to wait (to try to trigger overlaps) and use a new helper
utility function to make it more obvious what is happening
in the multilock release method.

Also swaps out the usage of time.time in the lock utils
unit test to attempt to use monotonic time (when it can
be used) so that false positives are unable to happen (using
time.time it is possible for time to go backwards and cause
a non-existent overlap to appear).

Change-Id: Ifb3967e1c1da41e1b7ac0793fc6a99f84de2a907