Updated openstack/openstack Project: openstack/neutron  8710c5ae92dc558c3b8a9026fd13f6663762b5eb
Skip DBDuplicateEntry exception in security group creation

An DBDuplicateEntry expection appeared if it is attempted to
create secutity groups in more then one thread. In this situation
at one moment it is trying to be created 2 default security groups
which is not allowed. In this case exception should skipped with
a message about failed attempt.

Closes-bug: #1419723

Change-Id: Id6fe997ad3b72f875160e1e31a3dc17c3cc53b75