Updated openstack/openstack Project: openstack/nova  af27a10272679bea37dbba6448388ea89fcc0853
VMware: fix resize of ephemeral disks

Ensure that a resize operation will work on ephemeral disks. If the
flavor or block device mapping require ephemerals then these will be
updated or created. The user is able to go from a flavor with or without
ephemerals to a flavor with or without ephemerals.
The resize will honor the configuration.

DocImpact - Resize of ephemeral disks no works for VMware driver.

Change-Id: I4b45299fdce59f4c52458d7515d0092b9f3b8db5