Updated openstack/openstack Project: openstack/tempest-lib  d7c3f6b232cb9a6fab8e41ed946473de0973db60
Summarize expected failures

The incoming stream has the xfail data but count_tests was not
matching on the xfail status. Instead it was matching on the
fragment 'fail' meaning that expected failures were counted as
such. By bounding the regular expressions it is possible to get more
specific results will still leaving the count_tests method flexible
for other users where fragments would be useful.

Counts of 'uxsuccess' are also summarized as "Unexpected Success".

It's unclear how to effectively automate testing of this. Manual
testing returns the expected results.

Change-Id: I5b1458f9a98712ea3e424d2c9610b915055138af