Updated openstack/openstack Project: openstack/cinder  6d58ec294aeadb23adc3903d357edbd5787ccf36
Dell sc driver iscsi multipath enhancement

Added support for the connector's multipath boolean.  If this is
set the initialize_connection function returns arrays of all the
IQNs, LUNs and portals associated with the mapping of the volume
to the server.  If this is not set the standard return of a lone
IQN, LUN and portal is implemented.

Minor change to find_volume.  If a volume name or instance isn't
set then it will not be sent as a payload filter to the Dell
Storage Center array.

Additional Unit Tests to support the changes.

Change-Id: I2f7404c76c71f5bf6d06496ab5aa4617ba80615b
Implements: blueprint dell-sc-driver-iscsi-multipath-enhancement