Updated openstack/openstack Project: openstack/cinder  8bda008dd0594f006e21d2745b616a97d7119065
Implement refresh as kwargs in get_volume_stats

When implementing the base driver "get_volume_stats" function, not all
drivers implemented 'refresh' as a kwarg. This can cause failures for any
code that simply invokes the following (assuming that refresh will default
to False -- which is does for 50+ drivers):

    driver.get_volume_stats()

This patch set updates the 7 drivers that did not implement refresh as a
kwarg so that the various driver implementations are consistent.

Closes-bug: 1423357
Change-Id: I6520155d1cbde4bf41573c170f4ea81815b5c1c2