Updated openstack/openstack Project: openstack/nova  50c97e2f16427b818adbeea60ce1da3cc6a7c7df
libvirt: cleanup setattr usage in test_host

There is some suspect setattr/delattr in test_host for mocking out
VIR_CONNECT_BASELINE_CPU_EXPAND_FEATURES. This can be easily stubbed out
with mock and the create=True kwarg so we can remove the setattr
ugliness.

Note that if you decorate a test method with a mock and use the new
kwarg, the stub isn't passed as an argument to the test method.

Change-Id: I414ac96b089f297c35fa821b926546290ac66bdc