Updated openstack/openstack Project: openstack/nova  c4a9cc56b2e783f5968c2696377ba8d488ed357d
Fix VNC access, when reverse DNS lookups fail

If DNS is configured on a node in a way, so that reverse DNS look ups
time out, noVNC will fail to connect to an instance with 'Connect
timeout' error.

The reverse DNS look up is done implicitly in BaseHTTPRequestHandler
(part of standard library), when logging a request. It's not
configurable, so the only way to disable it is to override the method
of the base class.

This is only true for the standard http server (used for novncproxy),
as the eventlet implementation (used for xvpvncproxy) seems to use
plain IP addresses without any reverse DNS look ups.

Closes-Bug: #1422315

Change-Id: I8b922de49134eda4d753a8e906b0de55203877d3