Updated openstack/openstack Project: openstack/nova  13223b5cc1b70f53beebeea87b7d85cfd694af20
Don't assume contents of values after aggregate_update

aggregate_update() was explicitly setting values['metadata'] for the
caller. However, it has also popped availability_zone, which it
doesn't replace. Also, other _update() functions in the db api don't
do this after modifying values in place, so anything which assumed
this behaviour would be inconsistent.

The only code which did assume this behaviour was in a test. This
patch fixes the test.

Change-Id: Iaa21576655da037e6186e900234e2c29c9e3a652