Updated openstack/openstack Project: openstack/neutron  34a44ae074fcd6b60519b06ce22a22b1b7d9869b
Remove error logs for a common situation (non created ipsets)

The log message was initially added by me as part of a
iptables_firewall refactor.

Ipsets for empty IP address lists aren't currently created,
that means that we can't reference empty security groups
(as ipsets) via iptable rules, and that's a normal condition,
not an error.

Closes bug: #1421772
Change-Id: I6b1ae1fb505ce5e76ef8cf7ef7df38cff57e0000