Updated openstack/openstack Project: openstack/keystone  05d8a6b82bf1f16c064bbee84d95c88b73030fae
Provide additional detail if OAuth headers are missing

If WSGIPassAuthorization is not set correctly, the user gets back
an obscure message, since the code continues and tries to get
values from a None object.

This change brings a proper exception for this configuration issue
and lets the user know what to fix.

Closes-Bug: #1420121

Change-Id: Iee282383a67ec13b5bfb5267fe9d81124c5e289d