Updated openstack/openstack Project: openstack/taskflow  83dbebc21150fc98a514cdf174cd11569deee341
Make the dispatcher handler be an actual type

Instead of having the dispatcher target be a tuple
or a single callback, have it be an actual type with
comments as to what the types fields are and how they
are used.

This makes it more obvious as to what those fields are
and how they are used so that it is easier to understand
how the WBE engine and components work.

Change-Id: I5541ccd5a7aa6ae73ed9adceeac2c0524e2a1dc9