Updated openstack/openstack Project: openstack/nova  8b4be1a2f9a16792ca1db595d6be5d10187096e4
XenAPI: Fix data loss on resize up

Resize up does not copy all data to the destination if data is written
between the start of the resize and the shutdown of the instance.

It seems we copy only the empty snapshot marker VDI instead of the VDI
that the VM keeps writing to during the resize operation. This fix
changes resize up behavior to copy the correct VDI.

Change-Id: I5eebdfaaa82481726ba6ce2a946773b6ea503905
Co-Authored-By: John Garbutt <john@johngarbutt.com>
Closes-Bug: #1421707