Updated openstack/openstack Project: openstack/neutron  ac65bca3baa10c58d5368dcf6474988f465fc8c9
Cisco Nexus ML2 Vendor decomposition

This adds a thin shim replacing the Cisco Nexus mech driver so that it
may use the code now stored in the networking-cisco stackforge. The code
for this driver that has been moved into the stackforge is also removed by
this patch.

Partial-Implements: blueprint core-vendor-decomposition
Change-Id: Ib88fd121509e96be42fb0bd22422ccaa7d4eb0f1
Closes-Bug: 1419010