Updated openstack/openstack Project: openstack/python-novaclient  be41ae238d05a2c3ade3822d869d2d199444c52a
add functional test for nova volume-attach bug

This is a functional test that boots a server via the cli, creates a
volume, and tries to attach it via the cli (which causes a failure due
to completion cache code). Note: the failure actually happens *after*
the attach command is dispatched, so the volume attach will still
work, the user is presented an error though.

Many TODOs remain for future patches. The test also tries to document
what was learned about the CLI redirection to cinder API, which was
introduced when Cinder was split out, but was tribal knowledge that
was lost in the mists of time.

Related-Bug: #1423695

Change-Id: Iaf474298be135843bff0114cf211bee19762f3ad