Updated openstack/openstack Project: openstack/keystone  26efa31c7bc2373f55d22a6bdcd4f473e5b5b5ad
Remove list_endpoint_groups_for_project from sample policies

The "identity:list_endpoint_groups_for_project" was defined in
the sample policy files but there's no actual mapping for it, so
setting a value for this target has no effect. If or when the
mapping gets added then this target must be added back in.

Change-Id: I1843bf61681944ae82c3325e0bb00c32ca8f1c16
Related-Bug: 1421968