Updated openstack/openstack Project: openstack/keystone  3ef58171ab1e52d693660417bd4699fabc72ed25
Add CADF notifications for most resources

Add a configuration option to define notification format for Keystone
resources (project, user, role, etc...) for specific actions,
such as: create/update/delete.

The end user has a choice, in the form of a config option to emit
either cadf notifications, or the previous (basic) format.

A seperate patch will include the refactoring needed in the manager
and controller classes to actually pass in the context, and get
the user's ID.

implements bp: cadf-everywhere

Change-Id: Ibb9efa4369627a92cf5779fedfc53132472fa7e6